modify folder/file names
