+++ src/components/Companies/Partners.jsx
