+++ src/components/Layouts/Footer/Footer.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import ButtonGroup from '../ui/buttons';
-import W2011 from '../../assets/img/previous-editions/2011W.svg';
-import W2012 from '../../assets/img/previous-editions/2012W.svg';
-import W2013 from '../../assets/img/previous-editions/2013W.svg';
-import W2014 from '../../assets/img/previous-editions/2014W.svg';
-import W2015 from '../../assets/img/previous-editions/2015W.svg';
-import W2016 from '../../assets/img/previous-editions/2016W.svg';
-import W2017 from '../../assets/img/previous-editions/2017W.svg';
+import ButtonGroup from '../../UI/Buttons';
+import W2011 from '../../../assets/img/previous-editions/2011W.svg';
+import W2012 from '../../../assets/img/previous-editions/2012W.svg';
+import W2013 from '../../../assets/img/previous-editions/2013W.svg';
+import W2014 from '../../../assets/img/previous-editions/2014W.svg';
+import W2015 from '../../../assets/img/previous-editions/2015W.svg';
+import W2016 from '../../../assets/img/previous-editions/2016W.svg';
+import W2017 from '../../../assets/img/previous-editions/2017W.svg';
 
 const Footer = () => {
   return (
