+++ src/components/Layouts/Footer/Footer.test.js
@@ -1,9 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Partners from './partners';
+import Footer from './Footer';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Partners />, div);
+  ReactDOM.render(
+    <Router>
+      <Footer />
+    </Router>,
+    div
+  );
   ReactDOM.unmountComponentAtNode(div);
 });
