+++ src/components/Practical/index.jsx
